refactor: move common usb components to separate module
