+++ src/lib.rs
@@ -19,6 +19,7 @@ pub mod hmacmode;
 mod manager;
 pub mod otpmode;
 mod sec;
+mod usb;
 
 use aes::cipher::generic_array::GenericArray;
 
@@ -27,10 +28,10 @@ use config::{Config, Slot};
 use configure::DeviceModeConfig;
 use error::ChallengeResponseError;
 use hmacmode::Hmac;
-use manager::{Flags, Frame};
 use otpmode::Aes128Block;
 use rusb::{Context, UsbContext};
 use sec::{crc16, CRC_RESIDUAL_OK};
+use usb::{Flags, Frame};
 
 const VENDOR_ID: [u16; 3] = [
     0x1050, // Yubico ( Yubikeys )
@@ -83,13 +84,13 @@ impl ChallengeResponse {
         let command = Command::DeviceSerial;
 
         let d = Frame::new(challenge, command); // FixMe: do not need a challange
-        let mut buf = [0; manager::STATUS_UPDATE_PAYLOAD_SIZE];
+        let mut buf = [0; usb::STATUS_UPDATE_PAYLOAD_SIZE];
         manager::wait(&mut handle, |f| !f.contains(Flags::SLOT_WRITE_FLAG), &mut buf)?;
 
         manager::write_frame(&mut handle, &d)?;
 
         // Read the response.
-        let mut response = [0; manager::RESPONSE_SIZE];
+        let mut response = [0; usb::RESPONSE_SIZE];
         manager::read_response(&mut handle, &mut response)?;
         manager::close_device(handle, interfaces)?;
 
@@ -204,7 +205,7 @@ impl ChallengeResponse {
 
     pub fn write_config(&mut self, conf: Config, device_config: &mut DeviceModeConfig) -> Result<()> {
         let d = device_config.to_frame(conf.command);
-        let mut buf = [0; manager::STATUS_UPDATE_PAYLOAD_SIZE];
+        let mut buf = [0; usb::STATUS_UPDATE_PAYLOAD_SIZE];
 
         let (mut handle, interfaces) =
             manager::open_device(&mut self.context, conf.device.bus_id, conf.device.address_id)?;
@@ -228,17 +229,17 @@ impl ChallengeResponse {
         let command = Command::DeviceSerial;
 
         let d = Frame::new(challenge, command); // FixMe: do not need a challange
-        let mut buf = [0; manager::STATUS_UPDATE_PAYLOAD_SIZE];
+        let mut buf = [0; usb::STATUS_UPDATE_PAYLOAD_SIZE];
         manager::wait(
             &mut handle,
-            |f| !f.contains(manager::Flags::SLOT_WRITE_FLAG),
+            |f| !f.contains(usb::Flags::SLOT_WRITE_FLAG),
             &mut buf,
         )?;
 
         manager::write_frame(&mut handle, &d)?;
 
         // Read the response.
-        let mut response = [0; manager::RESPONSE_SIZE];
+        let mut response = [0; usb::RESPONSE_SIZE];
         manager::read_response(&mut handle, &mut response)?;
         manager::close_device(handle, interfaces)?;
 
@@ -271,17 +272,17 @@ impl ChallengeResponse {
 
         (&mut challenge[..chall.len()]).copy_from_slice(chall);
         let d = Frame::new(challenge, command);
-        let mut buf = [0; manager::STATUS_UPDATE_PAYLOAD_SIZE];
+        let mut buf = [0; usb::STATUS_UPDATE_PAYLOAD_SIZE];
         manager::wait(
             &mut handle,
-            |f| !f.contains(manager::Flags::SLOT_WRITE_FLAG),
+            |f| !f.contains(usb::Flags::SLOT_WRITE_FLAG),
             &mut buf,
         )?;
 
         manager::write_frame(&mut handle, &d)?;
 
         // Read the response.
-        let mut response = [0; manager::RESPONSE_SIZE];
+        let mut response = [0; usb::RESPONSE_SIZE];
         manager::read_response(&mut handle, &mut response)?;
         manager::close_device(handle, interfaces)?;
 
@@ -312,17 +313,17 @@ impl ChallengeResponse {
 
         (&mut challenge[..chall.len()]).copy_from_slice(chall);
         let d = Frame::new(challenge, command);
-        let mut buf = [0; manager::STATUS_UPDATE_PAYLOAD_SIZE];
+        let mut buf = [0; usb::STATUS_UPDATE_PAYLOAD_SIZE];
 
         manager::wait(
             &mut handle,
-            |f| !f.contains(manager::Flags::SLOT_WRITE_FLAG),
+            |f| !f.contains(usb::Flags::SLOT_WRITE_FLAG),
             &mut buf,
         )?;
 
         manager::write_frame(&mut handle, &d)?;
 
-        let mut response = [0; manager::RESPONSE_SIZE];
+        let mut response = [0; usb::RESPONSE_SIZE];
         manager::read_response(&mut handle, &mut response)?;
         manager::close_device(handle, interfaces)?;
